chore(search-page): break up filters into separate files
